Ensure changes are done with the logged in user

+++ src/routes/tweetRoutes.ts
@@ -62,6 +62,21 @@ router.get('/:id', async (req, res) => {
 router.put('/:id', async (req, res) => {
 	const { id } = req.params;
 	const { content, image } = req.body;
+	// @ts-ignore
+	const user = req.user;
+
+	// determine whether the logged in user is updating one of their tweets
+	const tweet = await prisma.tweet.findUnique({
+		where: { id: id },
+		include: {
+			user: {
+				select: {
+					id: true,
+				},
+			},
+		},
+	});
+	if (user.id !== tweet?.user.id) return res.status(404).json({ error: 'You are not allowed to update this tweet' });
 
 	try {
 		const result = await prisma.tweet.update({
@@ -78,6 +93,21 @@ router.put('/:id', async (req, res) => {
 // delete one tweet
 router.delete('/:id', async (req, res) => {
 	const { id } = req.params;
+	// @ts-ignore
+	const user = req.user;
+
+	// determine whether the logged in user is deleting one of their tweets
+	const tweet = await prisma.tweet.findUnique({
+		where: { id: id },
+		include: {
+			user: {
+				select: {
+					id: true,
+				},
+			},
+		},
+	});
+	if (user.id !== tweet?.user.id) return res.status(404).json({ error: 'You are not allowed to delete this tweet' });
 
 	await prisma.tweet.delete({ where: { id: id } });
